+++ src/main/kotlin/uk/co/ben_gibson/git/link/git/Hosts.kt
@@ -1,6 +1,6 @@
 package uk.co.ben_gibson.git.link.git
 
-import git4idea.repo.GitRemote
+import java.net.URI
 import java.util.UUID
 
 class Hosts(private val hosts: Set<Host>) {
@@ -8,10 +8,9 @@ class Hosts(private val hosts: Set<Host>) {
     fun getById(id: String) = getById(UUID.fromString(id))
     fun getById(id: UUID) = hosts.firstOrNull() { it.id == id }
 
-    fun forRemote(remote: GitRemote) = remote.httpUrl()
-        ?.host
-        ?.let { host -> hosts.firstOrNull { it.baseUrl?.host == host } }
+    fun getByHostDomain(domain: URI) = hosts.firstOrNull { host -> host.domains.contains(domain) }
 
-    fun toSet() = hosts.toHashSet()
+    fun toSet() = hosts.toSet()
+    fun toList() = hosts.toList()
     fun toArray() = hosts.toTypedArray()
 }