+++ src/main/kotlin/uk/co/ben_gibson/git/link/pipeline/Middleware.kt
@@ -1,13 +1,11 @@
 package uk.co.ben_gibson.git.link.pipeline
 
-import com.intellij.openapi.project.Project
-import uk.co.ben_gibson.git.link.Context
-import java.net.URL
+import java.net.URI
 
 interface Middleware : Comparable<Middleware> {
     val priority: Int
 
-    operator fun invoke(pass: Pass, next: () -> URL?) : URL?
+    operator fun invoke(pass: Pass, next: () -> URI?) : URI?
 
     override fun compareTo(other: Middleware): Int {
         return priority - other.priority