+++ src/main/kotlin/uk/co/ben_gibson/git/link/settings/ApplicationSettings.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.XmlSerializerUtil
 import com.intellij.util.xmlb.annotations.Tag;
+import java.net.URI
 import java.util.UUID
 
 /**
@@ -14,13 +15,15 @@ import java.util.UUID
  */
 @State(name = "uk.co.ben_gibson.git.link.SettingsState", storages = [Storage("GitLink.xml")])
 class ApplicationSettings : PersistentStateComponent<ApplicationSettings?> {
-    private var listeners: MutableList<ChangeListener> = mutableListOf()
+    private var listeners: List<ChangeListener> = listOf()
 
-    var customHosts: MutableList<CustomHostSettings> = mutableListOf()
+    var customHosts: List<CustomHostSettings> = listOf()
         set(value) {
             field = value
             notifyListeners()
         }
+
+    var customHostDomains: Map<String, Set<String>> = mapOf()
 
     var lastVersion: String? = null
     var hits = 0
@@ -43,11 +46,16 @@ class ApplicationSettings : PersistentStateComponent<ApplicationSettings?> {
         var commitTemplate: String = ""
     )
 
+    fun findHostIdByCustomDomain(domain: URI) = customHostDomains
+        .entries
+        .firstOrNull { entry -> entry.value.contains(domain.toString()) }
+        ?.key
+
     fun registerListener(listener: ChangeListener) {
-        listeners.add(listener)
+        listeners = listeners.plus(listener)
     }
 
-    fun incrementHits() {
+    fun recordHit() {
         hits++
     }