+++ src/main/kotlin/uk/co/ben_gibson/git/link/url/factory/UrlFactoryLocator.kt
@@ -2,16 +2,14 @@ package uk.co.ben_gibson.git.link.url.factory
 
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
-import uk.co.ben_gibson.git.link.git.ChromiumHost
-import uk.co.ben_gibson.git.link.git.Host
-import uk.co.ben_gibson.git.link.git.TemplatedHost
+import uk.co.ben_gibson.git.link.git.*
 
 @Service
 class UrlFactoryLocator {
     fun locate(host: Host) : UrlFactory {
         return when(host) {
-            is TemplatedHost -> TemplatedUrlFactory(host.urlTemplate)
-            is ChromiumHost -> service<ChromiumUrlFactory>()
+            is Chromium -> service<ChromiumUrlFactory>()
+            else -> service<TemplatedUrlFactoryProvider>().forHost(host)
         }
     }
 }